requirements.txt
# Backend Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.7

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# File handling
python-multipart==0.0.6

# ML/Computer Vision (Railway optimized)
opencv-python-headless==4.8.1.78
mediapipe==0.10.7
numpy==1.24.3
scikit-image==0.21.0

# Image processing
Pillow==10.1.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
